Add files via upload
This new version of stellgap_soundwave_lagrng adds options for specifying the ion density and electron temperature profiles with respect to either the normalized toroidal flux, or the square-root of normalized toroidal flux. It also adjusts the adiabatic index, gamma to a value (1.35) that is more typical for modeling Alfven instabilities. More detailed descriptions are given in the comments near the beginning of the code.